PythonCode/requirements.txt
pyserial
firebase-admin
psycopg2-binary
streamlit
pandas
streamlit-autorefresh
requests